+++ front/src/components/Conversation/MessageAssistant/ForkButton.jsx
@@ -1,0 +1,13 @@
+import { GitFork } from "lucide-react";
+
+export default function ForkButton({ handleForkConversation }) {
+    return (
+        <button onClick={handleForkConversation} title="Fork conversation" className=" h-[22px] w-[22px] cursor-pointer disabled:opacity-40">
+            {" "}
+            <GitFork
+                className="opacity-20 group-hover:opacity-100 transition-opacity duration-300 h-[22px] w-[22px] cursor-pointer text-[#009EE0]"
+                alt="fork_icon"
+            />
+        </button>
+    );
+}